fix: adds all type related files into types folder
