+++ lib/gotrue.dart
@@ -1,15 +1,15 @@
 library gotrue;
 
-export 'src/auth_options.dart';
-export 'src/constants.dart' show AuthChangeEvent, InviteType;
-export 'src/cookie_options.dart';
-export 'src/auth_exception.dart';
+export 'src/types/auth_options.dart';
+export 'src/constants.dart' show AuthChangeEvent, InviteType, OtpType;
+export 'src/types/cookie_options.dart';
+export 'src/types/auth_exception.dart';
 export 'src/gotrue_admin_api.dart';
 export 'src/gotrue_client.dart';
-export 'src/auth_response.dart';
-export 'src/open_id_connect_credentials.dart';
-export 'src/provider.dart';
-export 'src/session.dart';
-export 'src/subscription.dart' show Subscription;
-export 'src/user.dart';
-export 'src/user_attributes.dart';
+export 'src/types/auth_response.dart';
+export 'src/types/open_id_connect_credentials.dart';
+export 'src/types/provider.dart';
+export 'src/types/session.dart';
+export 'src/types/subscription.dart' show Subscription;
+export 'src/types/user.dart';
+export 'src/types/user_attributes.dart';
