+++ lib/src/types/auth_exception.dart
