+++ lib/src/types/auth_options.dart
