+++ lib/src/types/cookie_options.dart
