+++ lib/src/types/user.dart
